@@ -21,78 +21,78 @@
   "author": "Gavin M. Roy",
   "license": "BSD-3-Clause",
   "devDependencies": {
-    "@babel/core": "^7.2.0",
+    "@babel/core": "^7.18.10",
+    "@babel/eslint-parser": "^7.16.5",
     "@babel/plugin-proposal-class-properties": "^7.2.3",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/preset-env": "^7.2.0",
     "@babel/preset-react": "^7.0.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.32",
-    "@fortawesome/free-brands-svg-icons": "^5.15.1",
-    "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "@fortawesome/react-fontawesome": "^0.1.14",
+    "@fortawesome/fontawesome-svg-core": "^1.3.0",
+    "@fortawesome/free-brands-svg-icons": "^5.15.4",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@headlessui/react": "^1.4.0",
     "@sentry/react": "^6.0.4",
-    "@tailwindcss/forms": "^0.2.1",
-    "@testing-library/dom": "^7.1.4",
+    "@tailwindcss/forms": "^0.5.2",
+    "@testing-library/dom": "^8.11.1",
     "@testing-library/jest-dom": "^5.3.0",
-    "@testing-library/react": "^11.2.3",
-    "@testing-library/user-event": "^12.6.0",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
     "@types/react": ">=16",
-    "autoprefixer": "^10.0.2",
-    "babel-eslint": "^10.1.0",
+    "autoprefixer": "^10.4.8",
     "babel-loader": "^8.0.4",
+    "babel-plugin-dynamic-import-webpack": "^1.1.0",
     "babel-plugin-transform-regenerator": "^6.26.0",
-    "babel-polyfill": "^6.26.0",
-    "browserslist": "^4.16.5",
+    "browserslist": "^4.19.1",
     "chart.js": "^3.3.2",
-    "css-loader": "^5.0.1",
+    "css-loader": "^6.5.1",
     "depcheck": "^1.3.1",
-    "dns-packet": "^1.3.2",
+    "dns-packet": "^5.3.0",
     "elliptic": "6.5.4",
-    "eslint": "^7.18.0",
-    "eslint-plugin-jest": "^24.1.3",
+    "eslint": "^8.5.0",
+    "eslint-plugin-jest": "^25.3.0",
     "eslint-plugin-react": "^7.11.1",
     "eslint-plugin-react-hooks": "^4.2.0",
     "fast-json-patch": "^3.0.0-1",
     "file-loader": "^6.2.0",
-    "hosted-git-info": "^2.8.9",
+    "hosted-git-info": "^4.0.2",
     "husky": "^7.0.0",
-    "i18next": "^19.3.4",
+    "i18next": "^21.6.3",
     "id128": "^1.6.3",
     "identity-obj-proxy": "^3.0.0",
-    "inter-ui": "^3.18.1",
-    "jest": "^26.6.3",
-    "jsonschema": "^1.4.0",
+    "jest": "^27.4.5",
+    "jsonschema": "^1.4.1",
     "lint-staged": ">=10",
     "lodash": "^4.17.21",
-    "luxon": "^1.27.0",
-    "node-fetch": "^2.6.1",
+    "luxon": "^3.0.1",
+    "node-fetch": "^3.1.0",
     "postcss": "^8.2.10",
     "postcss-import": "^14.0.0",
-    "postcss-loader": "^4.1.0",
+    "postcss-loader": "^6.2.1",
     "prettier": "^2.0.4",
     "prop-types": "^15.7.2",
-    "react": "17.0.1",
-    "react-chartjs-2": "^3.0.3",
+    "react": "17.0.2",
+    "react-chartjs-2": "^4.0.0",
     "react-circular-progressbar": "^2.0.3",
-    "react-dom": "17.0.1",
+    "react-dom": "17.0.2",
     "react-gravatar": "^2.6.3",
     "react-hot-loader": "^4.13.0",
     "react-i18next": "^11.3.4",
-    "react-markdown": "^6.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-select": "^4.3.0",
+    "react-markdown": "^6.0.3",
+    "react-router-dom": "^6.3.0",
+    "react-select": "^5.2.1",
     "react-sparklines": "^1.7.0",
     "react-syntax-highlighter": "^15.4.3",
-    "resolve-url-loader": "^3.1.2",
-    "slugify": "^1.4.6",
-    "ssri": "^6.0.2",
-    "style-loader": "^2.0.0",
-    "tailwindcss": "^2.0.1",
-    "typeface-inter": "^3.15.1",
+    "slugify": "^1.6.4",
+    "ssri": "^8.0.1",
+    "style-loader": "^3.3.1",
+    "tailwindcss": "^3.1.8",
     "typescript": ">= 3.7.0-beta",
-    "webpack": "^4.27.1",
+    "url": "^0.11.0",
+    "webpack": "^5.65.0",
     "webpack-cli": "^4.4.0",
-    "webpack-dev-server": "^3.11.2"
+    "webpack-dev-server": "^4.6.0",
+    "webpack-merge": "^5.8.0"
   },
   "husky": {
     "hooks": {
