@@ -9,8 +9,7 @@ class Backdrop extends React.PureComponent {
           className={
             (this.props.wait ? 'cursor-wait ' : '') +
             'flex items-end justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0'
-          }
-        >
+          }>
           <div className="fixed inset-0 transition-opacity" aria-hidden="true">
             <div className="absolute inset-0 bg-gray-500 opacity-75" />
           </div>
