@@ -27,8 +27,7 @@ class Badge extends React.PureComponent {
             (this.props.className !== undefined
               ? ` ${this.props.className}`
               : '')
-          }
-        >
+          }>
           {this.props.children}
         </a>
       )
@@ -39,8 +38,7 @@ class Badge extends React.PureComponent {
             colors[this.props.color]
           }` +
           (this.props.className !== undefined ? ` ${this.props.className}` : '')
-        }
-      >
+        }>
         {this.props.children}
       </div>
     )
