@@ -10,8 +10,7 @@ class Button extends React.PureComponent {
           className={
             !this.props.disabled ? this.props.className : 'btn-disabled'
           }
-          to={this.props.destination}
-        >
+          to={this.props.destination}>
           {this.props.children}
         </NavLink>
       )
@@ -25,8 +24,7 @@ class Button extends React.PureComponent {
             this.props.onClick(event)
           }
         }}
-        type={this.props.type}
-      >
+        type={this.props.type}>
         {this.props.children}
       </button>
     )
