@@ -24,7 +24,9 @@ class ErrorBoundary extends React.Component {
   }
 
   render() {
-    if (this.state.error !== null) return <Error>{this.state.error}</Error>
+    if (this.state.error !== null) {
+      return <Error>{`${this.state.error}`}</Error>
+    }
     return this.props.children
   }
 }
