@@ -77,8 +77,7 @@ function Select({
       }}
       placeholder={placeholder}
       ref={ref}
-      required={required}
-    >
+      required={required}>
       {multiple !== true && (
         <option value="">{placeholder !== undefined ? placeholder : ''}</option>
       )}
@@ -88,8 +87,7 @@ function Select({
             key={name + '-' + option.value}
             value={
               option.value !== null ? option.value.toString() : option.value
-            }
-          >
+            }>
             {option.label}
           </option>
         )
