@@ -21,8 +21,7 @@ function SimpleForm({
       onSubmit={(event) => {
         event.preventDefault()
         onSubmit()
-      }}
-    >
+      }}>
       {errorMessage !== null && (
         <Alert className="mb-3" level="error">
           {errorMessage}
@@ -33,15 +32,13 @@ function SimpleForm({
         <Button
           className={'btn-white'}
           disabled={saving === true}
-          onClick={onCancel}
-        >
+          onClick={onCancel}>
           {t('common.cancel')}
         </Button>
         <Button
           className={'btn-green'}
           disabled={ready === false || saving === true}
-          type="submit"
-        >
+          type="submit">
           {saving
             ? submitSavingText
               ? submitSavingText
