@@ -13,8 +13,7 @@ function NewMenuItem({ value, to }) {
       <NavLink
         className="new-menu-link"
         key={to.replace(/\//gi, '_') + '-nav-item'}
-        to={to}
-      >
+        to={to}>
         {value}
       </NavLink>
     </Menu.Item>
@@ -32,16 +31,14 @@ function NewMenu({ user }) {
     <Menu as="div" className="flex-shrink mr-3">
       <Menu.Button
         as="button"
-        className="bg-blue-700 hover:bg-blue-600 border border-blue-200 text-white hover:text-white px-3 py-1.5 rounded-md mr-5"
-      >
+        className="bg-blue-700 hover:bg-blue-600 border border-blue-200 text-white hover:text-white px-3 py-1.5 rounded-md mr-5">
         <Icon icon="fas plus-circle" />
         <Icon icon="fas caret-down" className="ml-3" />
       </Menu.Button>
       <Menu.Items
         aria-labelledby="user-menu"
         aria-orientation="vertical"
-        className="origin-top-right absolute right-20 mt-1 w-48 rounded-md shadow-lg py-1 focus:outline-none bg-white ring-1 ring-gray-300 ring-opacity-5 z-40"
-      >
+        className="origin-top-right absolute right-20 mt-1 w-48 rounded-md shadow-lg py-1 focus:outline-none bg-white ring-1 ring-gray-300 ring-opacity-5 z-40">
         <NewMenuItem
           to="/ui/operations-log/create"
           value={t('headerNavItems.newOperationsLogEntry')}
