@@ -7,8 +7,7 @@ class Footer extends React.PureComponent {
   render() {
     return (
       <div
-        className={`mt-5 sm:mt-6 text-right border-t border-gray-400 pt-5 mt-5 space-x-3 ${this.props.className}`}
-      >
+        className={`mt-5 sm:mt-6 text-right border-t border-gray-400 pt-5 mt-5 space-x-3 ${this.props.className}`}>
         {this.props.children}
         {this.props.onClose !== undefined &&
           this.props.closeText !== undefined && (
