@@ -37,8 +37,7 @@ function Container({
         setCurrentPage: setCurrentPage,
         setPageSize: setPageSize,
         startPage: startPage
-      }}
-    >
+      }}>
       {children}
     </Context.Provider>
   )
