@@ -15,15 +15,13 @@ function Section({ name, open, children }) {
     <div className="space-y-1 mt-2">
       <button
         className="bg-white text-gray-600 hover:text-blue-700 group w-full flex items-center pr-2 pt-2 rounded-md focus:outline-none"
-        onClick={onClick}
-      >
+        onClick={onClick}>
         <svg
           className={
             state === true ? indicatorClasses + ' rotate-90' : indicatorClasses
           }
           viewBox="0 0 20 20"
-          aria-hidden="true"
-        >
+          aria-hidden="true">
           <path d="M6 6L14 10L6 14V6Z" fill="currentColor" />
         </svg>
         {name}
