@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import React, { Fragment, useContext, useEffect, useState } from 'react'
-import { Route, useParams } from 'react-router-dom'
+import { Outlet, Route, Routes, useParams } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import {
   buildStyles,
@@ -80,19 +80,16 @@ function ProjectPage({ project, factTypes, refresh }) {
           <div className="flex-1 flex justify-end">
             <div
               className="flex-shrink mr-2"
-              style={{ height: '60px', width: '60px' }}
-            >
+              style={{ height: '60px', width: '60px' }}>
               <Tooltip
                 value={t('project.projectHealthScore')}
-                arrowPosition="right"
-              >
+                arrowPosition="right">
                 <CircularProgressbarWithChildren
                   value={project.project_score}
                   styles={buildStyles({
                     pathColor: color,
                     trailColor: '#ccc'
-                  })}
-                >
+                  })}>
                   <div className="absolute text-gray-600 font-semibold text-lg">
                     {parseInt(project.project_score)}
                   </div>
@@ -106,8 +103,7 @@ function ProjectPage({ project, factTypes, refresh }) {
         </Markdown>
         <nav
           className="relative z-0 rounded-lg shadow flex divide-x divide-gray-200"
-          aria-label="Tabs"
-        >
+          aria-label="Tabs">
           <Tab to={baseURL} isFirst={true}>
             {t('common.overview')}
           </Tab>
@@ -125,37 +121,44 @@ function ProjectPage({ project, factTypes, refresh }) {
             </Tab>
           )}
         </nav>
-        <Fragment>
-          <Route path={`/ui/projects/${project.id}`} exact>
-            <Overview
-              factTypes={factTypes}
-              project={project}
-              refresh={refresh}
-              urlPath={baseURL}
-            />
-          </Route>
-          <Route path={`/ui/projects/${project.id}/configuration`}>
-            <Configuration urlPath={baseURL} />
-          </Route>
-          <Route path={`/ui/projects/${project.id}/dependencies`}>
-            <Dependencies urlPath={baseURL} />
-          </Route>
-          <Route path={`/ui/projects/${project.id}/fact-history`}>
-            <FactHistory urlPath={baseURL} />
-          </Route>
-          <Route path={`/ui/projects/${project.id}/logs`}>
-            <Logs urlPath={baseURL} />
-          </Route>
-          <Route path={`/ui/projects/${project.id}/notes`}>
-            <Notes urlPath={baseURL} />
-          </Route>
-          <Route path={`/ui/projects/${project.id}/operations-log`}>
-            <OpsLog urlPath={baseURL} />
-          </Route>
-          <Route path={`/ui/projects/${project.id}/settings`}>
-            <Settings project={project} refresh={refresh} urlPath={baseURL} />
-          </Route>
-        </Fragment>
+        <Routes>
+          <Route
+            path={''}
+            element={
+              <Overview
+                factTypes={factTypes}
+                project={project}
+                refresh={refresh}
+                urlPath={baseURL}
+              />
+            }
+          />
+          <Route
+            path={`configuration`}
+            element={<Configuration urlPath={baseURL} />}
+          />
+          <Route
+            path={`dependencies`}
+            element={<Dependencies urlPath={baseURL} />}
+          />
+          <Route
+            path={`fact-history`}
+            element={<FactHistory urlPath={baseURL} />}
+          />
+          <Route path={`logs`} element={<Logs urlPath={baseURL} />} />
+          <Route path={`notes`} element={<Notes urlPath={baseURL} />} />
+          <Route
+            path={`operations-log`}
+            element={<OpsLog urlPath={baseURL} />}
+          />
+          <Route
+            path={`settings`}
+            element={
+              <Settings project={project} refresh={refresh} urlPath={baseURL} />
+            }
+          />
+        </Routes>
+        <Outlet />
       </div>
     </ErrorBoundary>
   )
