@@ -86,8 +86,7 @@ function DataTable({
         const value = (currentPage - 1) * pageSize
         setOffset(value)
       }}
-      setPageSize={setPageSize}
-    >
+      setPageSize={setPageSize}>
       <Table
         columns={columns}
         data={data}
