@@ -1,4 +1,4 @@
-import DateTime from 'luxon/src/datetime'
+import { DateTime } from 'luxon'
 import React, { Fragment, useContext, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
@@ -90,13 +90,11 @@ function Settings() {
             label: t('user.settings.authenticationTokens.title')
           }
         ]}
-        sideBarTitle="Available Settings"
-      >
+        sideBarTitle="Available Settings">
         <Form.Section
           name="api"
           title={t('user.settings.authenticationTokens.title')}
-          firstSection={true}
-        >
+          firstSection={true}>
           <Fragment>
             <div className="flex items-center justify-between my-3">
               <div
@@ -112,8 +110,7 @@ function Settings() {
                   onClick={(event) => {
                     event.preventDefault()
                     setShowTokenForm(true)
-                  }}
-                >
+                  }}>
                   <Fragment>
                     <Icon className="mr-3" icon="fas plus-circle" />
                     {t('user.settings.authenticationTokens.buttonText')}
@@ -239,8 +236,7 @@ function Settings() {
                   const newToken = document.getElementById('new-token')
                   newToken.select()
                   document.execCommand('copy')
-                }}
-              >
+                }}>
                 <Icon icon="fas clipboard" />
               </button>
             </div>
@@ -251,8 +247,7 @@ function Settings() {
               onClick={(event) => {
                 event.preventDefault()
                 setNewAuthToken(null)
-              }}
-            >
+              }}>
               {t('common.done')}
             </Button>
           </Modal.Footer>
